recipes: use individual files for recipes for better modularity
I tested this change in my kbase-devel instance. We will see how it goes
in the continuous integration tonight. I think it will be fine.

I tested the reload feature (router.py and consume.py use that feature, and I have
not altered these 2 modules at all) by adding a recipe in
lib/assembly/recipes/foo-bar.lisp (this one is not in the git repository)
without restarting anything and then by submitting a job with
"arast run -r foo-bar --data 2". It worked with the job completing with success.

I also tested the command "arast avail -r" (requesteed by Fangfang)
to list available recipes. It still works, so these changes have not broken this
feature.

The only 2 files that import the recipes modules are consume.py and router.py,
so things should be fairly safe.

This patch uses recipe_path = lib/assembly/recipes (to which gets appended the
root path). So this code is portable and won't break any existing systems when
they are eventually upgraded to use this patch. As Chris said, it would be better to add the
recipe_path option in the compute server configuration file. However, it is not clear what
is the best way to get the path to the configuration file from within the Python module
recipes.py. The main reason why is that code elsewhere (router.py + consume.py) use
functions in the function recipes.py and these function don't take an argument to
pass the configuration. One way to fix this would be to add the function
recipes.init(...) that would take a configuration file option.